finished implementing delete set

+++ Backend/src/main/java/com/husyairi/ManaProgressAku/Repository/ActivitySetRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ActivitySetRepository extends JpaRepository<ActivitySet, Integer> {
+public interface ActivitySetRepository extends JpaRepository<ActivitySet, Long> {
     List<ActivitySet> findByActivityActivityID(String activityID);
 }
